IZPACK-12: Java 5 boxing / unboxing updates.
git-svn-id: https://svn.codehaus.org/izpack/izpack-src/trunk@2060 7d736ef5-cfd4-0310-9c9a-b52d5c14b761

+++ src/lib/com/izforge/izpack/PackFile.java
@@ -177,7 +177,7 @@ public class PackFile implements Serializable
     public void setPreviousPackFileRef(String previousPackId, Long offsetInPreviousPack)
     {
         this.previousPackId = previousPackId;
-        this.offsetInPreviousPack = offsetInPreviousPack.longValue();
+        this.offsetInPreviousPack = offsetInPreviousPack;
     }
 
     /** The target operating system constraints of this file */
